Wed, Jun 22, 2022, 8:58 AM +00:00

+++ pset2/readability/readability.c
@@ -20,9 +20,9 @@ int main(void)
     int sentences = count_sentences(text);
 
     // Average number of letters per 100 words
-    float L = letters / words * 100;
+    float L = (float) letters / (float) words * 100;
     // Average number of sentences per 100 words
-    float S = sentences / words * 100;
+    float S = (float) sentences / (float) words * 100;
 
     // Calculate readability based on the Coleman-Liau index
     float index = 0.0588 * L - 0.296 * S - 15.8;
